The new apostrophe:receive-email task accepts a filename containing an email message and successfully extracts media from it and adds them to the media library, if and only if the "To:" address is of this form:
sitename+email_guid@somedomain

The email_guid must match the email_guid for a user on the site or media is not accepted.

There's a migration to give existing users a randomly generated email_guid, which works, and also a Doctrine listener to automatically set a random email_guid for each user at insert time.

I'll be moving some currently project specific partials that make users aware of the feature into this plugin shortly.



git-svn-id: http://svn.apostrophenow.org/plugins/apostropheEmailPlugin/trunk@4610 c89fab08-2f9b-42bf-8176-bd3ad6bee547